requirements.txt
asgiref==3.9.1
django==5.2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
phonenumberslite==9.0.10
pillow==11.3.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing-extensions==4.14.1

cryptography~=45.0.5